Added support for searching plugins.  It proceeds as follows:
1. If a plugin spec starts with a '/' then use it as an absolute
   path. That's the old behavior and probably useful to keep.

2. If you set the NPM_PLUGIN_PATH variable at build time, or if you
   leave it at its default value (see CMakeLists.txt and
   ros/CMakeLists.txt), then the colon-separated paths given there
   will be searched first.

3. If the environment variable NPM_PLUGIN_PATH is set (again in
   colon-separated format), then those directories are searched
   (possibly in addition to the compile-time NPM_PLUGIN_PATH).

Notice that for each directory that is searched, the following file
names are tried in order, where SPEC denotes the string passed to
npm::Plugin::search().

1. libSPEC.so
2. SPEC.so
3. SPEC

That should hopefully cover all practical cases... on Mac OS X the
modules built using CMake do not have a .dylib extension (at least on
10.8.5 with CMake 2.8.12.2) so there seems to be no need to add that
to the filenames to try.